requirements.txt
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.24.0
billiard==3.3.0.23
celery==3.1.18
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0.2
django-braces==1.13.0
django-cors-headers==2.5.2
django-elasticsearch-dsl==0.5.0
django-elasticsearch-dsl-drf==0.19
django-ipware==2.1.0
django-nine==0.2.2
django-nose==1.4.6
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
django-user-agents==0.3.2
djangorestframework==3.9.2
djangorestframework-simplejwt==4.1.1
elasticsearch==6.4.0
elasticsearch-dsl==6.1.0
gunicorn==19.9.0
idna==2.8
ipaddress==1.0.22
kombu==3.0.37
mailgun==0.1.1
nexmo==2.3.0
ngrok==0.0.1
nose==1.3.7
oauthlib==3.0.1
Pillow==5.4.1
psycopg2==2.7.7
pyactiveresource==1.0.1
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
python-decouple==3.1
python-http-client==3.1.0
python3-openid==3.1.0
pytz==2018.9
PyYAML==5.1
requests==2.21.0
requests-oauthlib==1.2.0
sendgrid==3.6.5
sendgrid-django==4.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
ua-parser==0.8.0
urllib3==1.24.1
user-agents==2.0
whitenoise==4.1.2
